backend/requirements.txt
fastapi
uvicorn[standard]
jinja2
aiofiles
psycopg2
sqlalchemy
python-dotenv
pydantic[email]
passlib[bcrypt]
pytest
requests
python-jose
python-multipart